[OFFER] Changed get requests to getting Optional
[ANDROID]
Changed project structure

+++ backend_spring/src/main/java/pl/pollub/inzynierka/domain/repositories/OfferRepository.java
@@ -13,6 +13,6 @@ public interface OfferRepository extends JpaRepository<OfferEntity, Long> {
     List<OfferDto> getAllOffers();
 
     @Query("select new pl.pollub.inzynierka.infrastructure.Offer.OfferDto(o.title,o.description,o.image,o.oldPrice,o.newPrize,o.gps,o.startDate,o.expireDate,o.isAvailable, o.categoryByCategoryId.id, o.id) from OfferEntity o where o.id = :id")
-    OfferDto getOfferById(int id);
+    OfferDto getOfferById(Long id);
 
 }
